🔨 Improve file structure and naming
