+++ src/SFA.DAS.ApprenticeCommitments/Data/Models/Apprenticeship.cs
@@ -23,10 +23,11 @@ namespace SFA.DAS.ApprenticeCommitments.Data.Models
         }
 
         public long Id { get; private set; }
-        public long CommitmentsApprenticeshipId { get; set; }
+        public long CommitmentsApprenticeshipId { get; private set; }
+        public Apprentice Apprentice { get; private set; }
+
         public ApprenticeshipDetails Details { get; private set; }
 
-        public Apprentice Apprentice { get; private set; }
         public bool? TrainingProviderCorrect { get; private set; }
         public bool? EmployerCorrect { get; private set; }
