+++ src/SFA.DAS.ApprenticeCommitments/Data/Models/ApprenticeshipDetails.cs
@@ -1,0 +1,28 @@
+﻿using SFA.DAS.ApprenticeCommitments.Exceptions;
+using System;
+using System.ComponentModel.DataAnnotations.Schema;
+using System.Net.Mail;
+
+#nullable enable
+
+namespace SFA.DAS.ApprenticeCommitments.Data.Models
+{
+    public class ApprenticeshipDetails
+    {
+        public ApprenticeshipDetails(
+            long employerAccountLegalEntityId, string employerName,
+            long trainingProviderId, string trainingProviderName)
+        {
+            EmployerAccountLegalEntityId = employerAccountLegalEntityId;
+            EmployerName = employerName;
+            TrainingProviderId = trainingProviderId;
+            TrainingProviderName = trainingProviderName;
+        }
+
+        public long EmployerAccountLegalEntityId { get; private set; }
+        public string EmployerName { get; private set; }
+
+        public long TrainingProviderId { get; private set; }
+        public string TrainingProviderName { get; private set; }
+    }
+}
